fix(issues): resolve u32/u64 group_first_seen bug (#7397)
I recently set up ingestion for group_first_seen. This column is defined
as `DateTime`.

When I set up the rust processor, I was basing my processing on the
`datetime` field, since I knew that the two were sent over identically
from `sentry/**/eventstream/snuba.py`. This included using the
`StringToIntDatetime64` deserializer, which I still believe is correct.
Unfortunately I followed that 64 further than I should've and set the
field to be a `u64` in the row — when it should've been a `u32`.

This PR fixes that to make future ingestions correct. (I based my
changes on the `timestamp` field's parsing, which is similarly a
`DateTime` (no 64) column.)