feat(deletes): add stub RPC endpoint for deleting all items in a trace (#7416)
For deleting PII in EAP, we aim to start with an approach that mirrors
what we do for search_issues. We want the request to arrive through RPC,
though. This change creates a stub for allowing a caller to request
deleting items for a trace, though it currently does nothing.

I'm separating this out to make the change that involves real
functionality smaller.

Requires https://github.com/getsentry/sentry-protos/pull/144

https://linear.app/getsentry/issue/EAP-282/create-new-rpc-endpoint-for-deletion-from-eap-items