ref(outcomes-routing): always downsample data older than 30 days (#7424)
Soon we will be only storing up to 30 days of data in the non-sampled
table (aka tier 1). When that is the case we want to make sure that any
queries that query for data that is older than 30 days will not route to
tier 1.

Added `enforce_unsampled_retention` runtime config so that we can test
this out and because we don't expect to enforce this right away.