feat: TraceItemStats heatmap, basic functionality (#7433)
please see here for the design doc of what this new endpoint is
https://www.notion.so/sentry/Design-Doc-Support-heat-map-representation-in-TraceItemStats-2628b10e4b5d80c1a0bbc4c1e7216783

this PR implements the most basic and essential functionality outlined
in the design doc. That is, heatmaps with string attribute on the x-axis
and numeric attributes on the y-axis. This enables users to create
heatmaps comparing the correlation of attributes such as `span.op
(string)` and `span.duration (number)`.

the next steps are the tests marked by `@pytest.mark.skip`, that is:
* numeric on x-axis, string on y (optional, since i could just swap
numeric onto y and string on x)
* string on both axis
* numeric on both axis