feat(health-check): add limit to individual health check cluster wait (#7436)
During a recent incident, the liveness check was exceeding 30s
regularly, probably due to a single cluster being unreachable. This
created a lot of flapping in the API, and so we want to minimize the
amount that any individual cluster can affect the `/health` endpoint
performance.

In this change, we use a thread pool executor and `as_completed()` to
limit concurrent query node health checks to 500ms. This is more than
p99 for how long the `/health` endpoint usually takes to respond.
Additionally, I've added run-time config to make it 2s temporarily in
our busiest environments