fix(eap): Set a better default for timestamp field and customize for logs (#7441)
For logs, `sentry.start_timestamp_precise` doesn't exist since it's a
span only attribute so far.

This is a workaround to use a better default for timestamp and customize
this for spans. A better fix would be to promote this field to a new
column allowing more precision and using this in conjunction with the
`timestamp` column since it's in the sort key.

---------

Co-authored-by: William Mak <william@wmak.io>