feat: add request_id to sentry error context (#7469)
recently I was trying to debug some of the query exceptions we have in
our sentry issues like this one
https://sentry.sentry.io/issues/6895313867/?project=300688&query=http.referer%3Aapi.dashboards.widget.line-chart.find-topn&referrer=issue-stream&seerDrawer=true
but I noticed that the request_id for the request that caused this issue
is nowhere in the issue, making it very hard to go to the snuba querylog
and find the corresponding erroring query. This PR adds the request_id
to the context when an error happens while querying clickhouse, this
should make it quick and easy to find the corresponding entry in the
querylog for an issue.