chore: increase max timeseries buckets (#7482)
We're migrating transaction alerts to spans. Transaction alerts
supported 1 min intervals so we need to be able to plot a 1 week query
with this interval
([example](https://sentry.sentry.io/monitors/2809118/?statsPeriod=7d)).
We don't intend to open this option up to new span alerts though. This
is just to ensure old transaction alerts can be migrated to spans
correctly.

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>