ref(deletes): default to bulk deletes (#7447)
**context:**
Eventually we'll just use the RPC endpoint added in
https://github.com/getsentry/snuba/pull/7426 (and earlier PRs) but as
for right now, we only want to allow bulk deletes. We don't need
`use_bulk_deletes` runtime config anymore so it's removed and we only
use `bulk_delete_from_storage`.

We can clean up the logic in `bulk_delete_query.py` and
`delete_query.py` after this as well as they share logic but we only
relaly should have the bulk delete logic

**note:**
the auto formatter re-formatted a lot of this so many of the changes are
formatting