[gocd] split snuba deploys into snuba-deploy-py and snuba-deploy-rs (#7507)
Split out the current deploy pipelines into `deploy-snuba-py` and
`deploy-snuba-rs`. The goal of this is to set us up for a future where
we can only deploy rust consumers when changes are made in the Rust path
(reducing the amount of SLO breaches on high-traffic consumers for
rebalances).

At this stage, we'll still deploy both with any code change to test.
Next, we'll want to change the triggering logic to ignore code changes
from irrelevant sources