Share allocation policies among routing strategies (#7499)
Closes:
https://linear.app/getsentry/issue/EAP-79/move-eap-allocation-policies-into-the-outcomes-routing-strategy


Currently, the flex time routing strategy and the regular outcomes based
strategy have dfiferent allocation policies (and the flex time one is
not even enabled).

Unless there is a very good reason, the allocation policies should be
the same and should thus share the same resource identifier.

There was also a missed cleanup in the allocation policy consolidation
(the autcomplete storage is still using storage allocation policies
unnecessarily), I cleaned that up

Tested by:

1. Spinning up snuba admin locally
2. making sure the default config override was correct
3. making sure changing the allocation policy config in one strategy
changed it for the other one