ref(lw-deletes): add runtime config lightweight_deletes_sync (#7522)
Adds runtime config option to change the
[`lightweight_deletes_sync`](https://clickhouse.com/docs/operations/settings/settings#lightweight_deletes_sync)
option for delete queries. We've seen that these `DELETE FROM ... ON
CLUSTER ...` take quite a long time and tend to end up being execute
async regardless.

The default for the setting is `2` which is the default I set for the
config option. I'm going to be testing the options in S4S first, but
eventually we should just hardcode the setting.