gocd: split pipelines, leaving original, update migration check script (#7525)
When deploying https://github.com/getsentry/snuba/pull/7510, we
encountered a couple problems.

First was that we were checking the health of a pipeline that we deleted
(snuba was removed to be replaced with snuba-rs and snuba-py). This was
addressed in https://github.com/getsentry/snuba/pull/7523 which made
snuba-py and snuba-rs check the s4s deploys for those pipelines,
respectively.

We then found, however, that `checks` was checking if every stage in a
pipeline had executed successfully. We could force bypass this to make a
deploy happen, but there was a circular dependency here where `checks`
could only succeed once `checks` had succeeded. So here I've changed
`scripts/fetch_service_refs.py` to make `checks` succeed as long as we
reach the `pipeline-complete` stage of a region successfully. So the
first deploy will have to manually bypass `checks` still but future
deploys should proceed unimpeded.

In order to reduce gocd-change-deploy thrash, I'm reintroducing the
original pipeline (#7510 deleted it while standing up the new ones),
which I'm hoping we can just leave paused as we work out remaining
kinks. But it eliminates the need to revert if we see any problems

Once we're deploying successfully on the new pipelines we can delete the
files:
- gocd/templates/pipelines/snuba.libsonnet
- gocd/templates/bash/deploy-st.sh
- gocd/templates/bash/deploy.sh