chore: add metric to see how many GetTrace request have a user provided limit (#7521)
I am beginning to rollout GetTrace endpoint pagination, [see this PR.
](https://github.com/getsentry/snuba/pull/7508) Right now the user must
opt-in to pagination by providing a limit. Eventually we would like to
enforce a global max limit that is always enforced, but before we can do
that we need all of our users to be ready to support pagination. I.e.
they need to be able to handle receiving the results in multiple pages.

I will tell product to add this support, and then begin sending a
`limit` in their request. Once all user queries contain a `limit`, I can
enable global max.