feat(deletes): support TraceItemFilter[s] in DeleteTraceItems API (#7535)
Wire DeleteTraceEndpoint TraceItemFilter[s] into bulk_delete_query
arguments. This allows product to test API-level integration with the
delete endpoint, without blowing up the size of this change.

This validates the incoming fields and passes them through in a new
`AttributeConditions` object when not deleting by trace_id

Since we don't support delete by attributes, and it's unsafe to only use
the column values provided, we short-circuit requests that have
attribute conditions (and do no delete at all).

Next steps:
1. Wire these values into the DeleteQueryMessage (undecided at the
moment if we should resolve the attributes to columns when we create
that message, or if attribute_conditions should be a new field that gets
resolved when the consumer executes the delete)
2. Run a test?

Depends on https://github.com/getsentry/snuba/pull/7534