spike: delegate to storage to do translation mapping (#7585)
Try to get rid of specialized eap_items bucketing logic in
snuba/web/delete_query in #7570

I think this is a good compromise between running a full query pipeline
and duplicating bucketing logic in delete_query. It will continue to
execute column translations as configuration is changed in the
entity/storage YAMLs, increasing decoupling but doesn't bring in a ton
of extra scope.

Why not run the entire query pipeline? Unfortunately, this would require
coupling all deletes to *entity* definitions (these mappers exist on the
entity), but we configure and pass around delete queries based on the
storage.

Tested end-to-end with this flow:
https://www.notion.so/sentry/e2e-attribute-testing-2ab8b10e4b5d80cea5cfdc255cff5305