feat(deletes): add EAP deletes allowlist for launch (#7591)
With this change, we see if an incoming delete is for an org ID that's
part of our allowlist. If not, we bypass execution, emit a metric, and
commit the offset.

If allowlist is not present, all deletes are executed as normal.

Setting this up so we can test for launch on specific orgs without
dialing up globally