ref(admin): allow system queries on nodes sans clusters (#7605)
Currently adding new replicas, shards, query nodes, etc is not something
that is easily done via snuba admin because of how we populate the
possible nodes for system queries. The are assumptions being made that
aren't necessarily true for these kinds of workflows:
1. **That a node is always part of a cluster**
1. in the case of a adding a query node, we want to be able to run the
create table queries on the node _before_ we add it to the query cluster
definition, otherwise traffic will be routing to a node that has no
tables.
   
2. **That a node in a storage cluster can be discoverable from the query
node**
1. in the case that we add a new replica we want to wait until tables
are created and replication on the new replica before the query nodes is
aware of the new replica
2. in the case that we add a new shard we also want to wait until all
tables are created on the new shard replicas before the query node can
route inserts to the new shard

This PR introduces the concept of running queries on `clusterless`
nodes, which includes nodes that are part of a storage cluster but not a
query cluster. Running these queries can be regular system queries or
sudo queries. The main difference is minimal validation since we don't
have cluster info to work off of. Instead of a drop down of nodes, you
input the IP address (which one can look up in GCP once the node has
been provisioned)

Below shows the UI changes


https://github.com/user-attachments/assets/69f04829-9510-4216-900e-d9db977deccc